+++ src/test/scala/edu/ie3/mobsim/model/builder/EvBuilderFromRandomModelSpec.scala
@@ -9,8 +9,7 @@ package edu.ie3.mobsim.model.builder
 import edu.ie3.mobsim.io.probabilities.ProbabilityDensityFunction
 import edu.ie3.mobsim.model.{ElectricVehicle, TripSimulationTestData}
 import edu.ie3.test.common.UnitSpec
-import tech.units.indriya.quantity.Quantities
-import tech.units.indriya.unit.Units
+import squants.Meters
 
 class EvBuilderFromRandomModelSpec
     extends UnitSpec
@@ -141,7 +140,7 @@ class EvBuilderFromRandomModelSpec
         givenHomePoi.copy(
           id = s"home_poi_$cnt",
           nearestChargingStations = Map(
-            givenChargingStation -> Quantities.getQuantity(1.0, Units.METRE)
+            givenChargingStation -> Meters(1.0)
           )
         )
       } ++ Range(100, 200).map { cnt =>